[refactor] AuthorViewParts 抽出と timeline_runtime_support の二分割 (WP-H5 PR2)
- PostView の作者まわり 8 フィールド(プロフィール由来 4 + フォロー
  関係 4)の導出ロジックを AuthorViewParts に集約し、5 箇所の重複を
  置換(repost 元の source_author_* 4 フィールドも同じ部品から充当)。
  フィールド追加・導出変更が 1 箇所修正になる
- timeline_runtime_support.rs(990 行)を購読・復旧
  (timeline_subscription_support.rs)と View 変換
  (timeline_view_support.rs)へ二分割

挙動不変(IPC 形状 golden = views_wire_snapshot 含む全テスト green)。
